app/requirements.txt
Flask~=2.2.3
Flask_mongoengine
Flask_APScheduler
python-dotenv~=1.0.0
requests~=2.28.2
praw~=7.7.0
beautifulsoup4~=4.12.2
mongoengine~=0.27.0
APScheduler~=3.10.1